+++ apps/remnant2toolkit/app/(builds)/base-game-builds/page.tsx
@@ -45,7 +45,7 @@ export default async function Page() {
           subtitle={
             <div className="flex flex-col">
               <BaseText>{NAV_ITEMS.baseGameBuilds.description}</BaseText>
-              <BaseTextLink href="/community-builds?releases=base&withQuality=true">
+              <BaseTextLink href="/community-builds?releases=base">
                 <span className="text-primary-500">
                   Want more? Click here to browse all community submitted base
                   game builds.