+++ apps/remnant2toolkit/app/(builds)/community-builds/_components/community-builds-list.tsx
@@ -89,6 +89,7 @@ export function CommunityBuildsList({
       <BuildList
         currentPage={currentPage}
         isLoading={isLoading}
+        isWithQuality={buildListFilters.withQuality}
         pageNumbers={pageNumbers}
         totalItems={totalBuildCount}
         totalPages={totalPages}