+++ apps/remnant2toolkit/app/(builds)/featured-builds/_components/featured-builds-list.tsx
@@ -89,6 +89,7 @@ export function FeaturedBuildsList({
       <BuildList
         currentPage={currentPage}
         isLoading={isLoading}
+        isWithQuality={buildListFilters.withQuality}
         pageNumbers={pageNumbers}
         totalItems={totalBuildCount}
         totalPages={totalPages}