+++ CHANGELOG.md
@@ -11,6 +11,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - Build variants are now live. These work similar to how Linked Builds work, with some notable differences. Namely, when visiting a build URL, the build variants will have tabs along the top showing the variants inline, without requiring clicks to a different page. Existing builds can have variants added to them. Additionally, creators can now add build variants while creating a build, making the process more seamless overall.
   - This is a new feature, and a big overhaul, so there may be some bugs. Please let me know in the Discord if you find any.
 
+### Changed
+
+- Quality Build filter is now on by default when viewing build feeds. There is a button above and below the build list to easily apply/remove the filter.
+
 ### Removed
 
 - Linked Builds are being deprecated. You can still view existing Linked Builds, but you can no longer add or edit them. Truthfully, I should have incorporated Linked Builds in the way that Build Variants now work at the outset. However, trying to manage the Linked Build architecture with the existence of Build Variants added too many odd scenarios that would add a lot of overhead moving forward that didn't feel worth it for the benefit that Build Variants provide compared to Linked Builds